# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/firm/{firmId}": {
      "get": {
        "description": "Retrieve all users belonging to a specific firm",
        "operationId": "UsersController_getUsersByFirm",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (required for all endpoints)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "client",
                "lawyer",
                "intern",
                "firm",
                "firmadmin",
                "superadmin",
                "user"
              ]
            }
          },
          {
            "name": "firmId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users for the firm",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Firm not found or has no users"
          }
        },
        "summary": "Get users by firm",
        "tags": [
          "users"
        ]
      }
    },
    "/users/login": {
      "post": {
        "description": "Authenticate using email and password.",
        "operationId": "UsersController_login",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (required for all endpoints)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "client",
                "lawyer",
                "intern",
                "firm",
                "firmadmin",
                "superadmin",
                "user"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password"
          }
        },
        "summary": "Login user",
        "tags": [
          "users"
        ]
      }
    },
    "/users": {
      "post": {
        "description": "Creates a new user with the specified role (client, lawyer, intern, firm, firmadmin, superadmin). Only firmadmins/superadmins can create users.",
        "operationId": "UsersController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (required for all endpoints)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "client",
                "lawyer",
                "intern",
                "firm",
                "firmadmin",
                "superadmin",
                "user"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "403": {
            "description": "Forbidden - only firmadmin role can create users"
          }
        },
        "summary": "Create a new user",
        "tags": [
          "users"
        ]
      },
      "get": {
        "description": "Retrieve all users or filter by role. Only firmadmins and superadmins can view all users.",
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (required for all endpoints)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "client",
                "lawyer",
                "intern",
                "firm",
                "firmadmin",
                "superadmin",
                "user"
              ]
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filter users by role",
            "schema": {
              "enum": [
                "client",
                "lawyer",
                "intern",
                "firm",
                "firmadmin",
                "superadmin",
                "user"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - only firmadmin role can view all users"
          }
        },
        "summary": "Get all users",
        "tags": [
          "users"
        ]
      }
    },
    "/users/firms/all": {
      "get": {
        "description": "Retrieve a list of all business entities (Firms).",
        "operationId": "UsersController_getAllFirms",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (required for all endpoints)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "client",
                "lawyer",
                "intern",
                "firm",
                "firmadmin",
                "superadmin",
                "user"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of firms"
          }
        },
        "summary": "Get all registered law firms",
        "tags": [
          "users"
        ]
      }
    },
    "/users/firms": {
      "post": {
        "description": "Adds a law firm to the platform without the 3-step self-service onboarding flow. If adminName, adminEmail and adminPassword are all supplied, the firm admin account is provisioned and linked to the new firm in the same call. A subscription on the chosen tier starts billing immediately.",
        "operationId": "UsersController_createFirm",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (required for all endpoints)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "client",
                "lawyer",
                "intern",
                "firm",
                "firmadmin",
                "superadmin",
                "user"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFirmDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Firm created successfully"
          },
          "403": {
            "description": "Forbidden - superadmin only"
          },
          "409": {
            "description": "Firm email or admin email already in use"
          }
        },
        "summary": "Create a law firm (superadmin only)",
        "tags": [
          "users"
        ]
      }
    },
    "/users/firms/{firmId}": {
      "put": {
        "description": "Patch any editable field on a law firm, including its pricing tier.",
        "operationId": "UsersController_updateFirm",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (required for all endpoints)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "client",
                "lawyer",
                "intern",
                "firm",
                "firmadmin",
                "superadmin",
                "user"
              ]
            }
          },
          {
            "name": "firmId",
            "required": true,
            "in": "path",
            "description": "Law firm ID",
            "schema": {
              "example": "firm-1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFirmDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Firm updated successfully"
          },
          "403": {
            "description": "Forbidden - superadmin only"
          },
          "404": {
            "description": "Firm not found"
          }
        },
        "summary": "Update a law firm (superadmin only)",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "description": "Removes a law firm from the platform. A firm that still has members is rejected so its users are never orphaned — pass cascade=true to delete those members too.",
        "operationId": "UsersController_deleteFirm",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (required for all endpoints)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "client",
                "lawyer",
                "intern",
                "firm",
                "firmadmin",
                "superadmin",
                "user"
              ]
            }
          },
          {
            "name": "firmId",
            "required": true,
            "in": "path",
            "description": "Law firm ID",
            "schema": {
              "example": "firm-1",
              "type": "string"
            }
          },
          {
            "name": "cascade",
            "required": false,
            "in": "query",
            "description": "Set to \"true\" to also delete every user belonging to this firm",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Firm deleted successfully"
          },
          "400": {
            "description": "Firm still has members and cascade was not set"
          },
          "403": {
            "description": "Forbidden - superadmin only"
          },
          "404": {
            "description": "Firm not found"
          }
        },
        "summary": "Delete a law firm (superadmin only)",
        "tags": [
          "users"
        ]
      }
    },
    "/users/firms/{firmId}/tier": {
      "put": {
        "description": "Change the pricing tier (Starter, Growth, Enterprise) of any law firm at any time.",
        "operationId": "UsersController_updateFirmTier",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (required for all endpoints)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "client",
                "lawyer",
                "intern",
                "firm",
                "firmadmin",
                "superadmin",
                "user"
              ]
            }
          },
          {
            "name": "firmId",
            "required": true,
            "in": "path",
            "description": "Law firm ID",
            "schema": {
              "example": "firm-1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tier updated successfully"
          },
          "400": {
            "description": "Invalid tier value"
          },
          "403": {
            "description": "Forbidden - superadmin only"
          },
          "404": {
            "description": "Firm not found"
          }
        },
        "summary": "Update a firm tier (superadmin only)",
        "tags": [
          "users"
        ]
      }
    },
    "/users/lawyers": {
      "get": {
        "description": "Returns all lawyers belonging to the given firm. firmadmin must pass their firmId via x-firm-id header — only lawyers with a matching firmId are returned. superadmin can omit the header to get all lawyers.",
        "operationId": "UsersController_getLawyers",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (required for all endpoints)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "client",
                "lawyer",
                "intern",
                "firm",
                "firmadmin",
                "superadmin",
                "user"
              ]
            }
          },
          {
            "name": "firmId",
            "required": false,
            "in": "query",
            "description": "Alternative: pass firmId as query param instead of header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-firm-id",
            "in": "header",
            "description": "The firm ID of the calling firm admin (e.g. \"firm-1\"). Omit to get all (superadmin only).",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of lawyers in the specified firm",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden – insufficient role"
          }
        },
        "summary": "Get lawyers for a firm (firmadmin / superadmin)",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Retrieve a specific user by their ID. Available to all roles.",
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (required for all endpoints)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "client",
                "lawyer",
                "intern",
                "firm",
                "firmadmin",
                "superadmin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - role header is required"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Get user by ID (all roles)",
        "tags": [
          "users"
        ]
      },
      "put": {
        "description": "Update user details like role, status, and availability",
        "operationId": "UsersController_updateUser",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (required for all endpoints)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "client",
                "lawyer",
                "intern",
                "firm",
                "firmadmin",
                "superadmin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data or email already registered"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Update a user",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "description": "Remove a user from the system",
        "operationId": "UsersController_deleteUser",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (required for all endpoints)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "client",
                "lawyer",
                "intern",
                "firm",
                "firmadmin",
                "superadmin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          }
        },
        "summary": "Delete a user",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "LexFlow API",
    "description": "REST API for the LexFlow Legal Platform.\n\n**RBAC:** Pass `role` in the request header to identify the caller.\nAccepted values: `client` | `lawyer` | `intern` | `firmadmin` | `superadmin`",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role-auth": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "User role for RBAC. Values: client | lawyer | intern | firmadmin | superadmin"
      }
    },
    "schemas": {
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "firmName": {
            "type": "string"
          },
          "fullName": {
            "type": "string",
            "description": "Full name of the user",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "john.doe@example.com"
          },
          "role": {
            "type": "string",
            "description": "User role",
            "enum": [
              "client",
              "lawyer",
              "intern",
              "firm",
              "firmadmin",
              "superadmin",
              "user"
            ],
            "example": "client"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Account creation timestamp",
            "example": "2026-04-30T10:30:00Z"
          },
          "phone": {
            "type": "string",
            "description": "Phone number"
          },
          "firmId": {
            "type": "string",
            "description": "Associated firm ID (for FIRMADMIN, LAWYER, INTERN roles)",
            "example": "firm-1"
          },
          "accountStatus": {
            "type": "string",
            "example": "active"
          },
          "availability": {
            "type": "string",
            "example": "available"
          }
        },
        "required": [
          "id",
          "fullName",
          "email",
          "role",
          "createdAt"
        ]
      },
...
    }
  }
}